FlexiSAF Internship .Big Data Project
This project explores the fundamentals of Big Data and Data Engineering. It covers the characteristics of Big Data, the role of data engineers, and the tools and technologies (e.g., Hadoop, Spark, SQL/NoSQL, ETL tools) used to process and analyze massive datasets. It highlights the importance of building scalable data pipelines and architectures to handle real-time and batch data efficiently.